moved axios calls
